Rename parameters to definitions
