+++ arlunio/__init__.py
@@ -1,10 +1,8 @@
 from ._color import RGB8  # noqa: F401
-from ._core import Collection  # noqa: F401
+from ._core import Collection, Definition, definition  # noqa: F401
 from ._expressions import all, any, invert  # noqa: F401
 from ._image import Image, Resolutions  # noqa: F401
 from ._loaders import load_parameters
-from ._parameters import Parameters as P  # noqa: F401
-from ._parameters import parameter  # noqa: F401
 from ._shapes import Canvas, Shape, ShapeCollection, load_shapes, shape  # noqa: F401
 from ._version import __version__  # noqa: F401
